Put back first argument of PPIDD_Initialize() to int so that argc can be passed directly from main.

+++ src/ppidd_c.h
@@ -1,6 +1,6 @@
 #ifndef __PPIDD_C_H__
 #define __PPIDD_C_H__
-   extern void PPIDD_Initialize(int64_t argc, char **argv);
+   extern void PPIDD_Initialize(int argc, char **argv);
    extern void PPIDD_Initialize_data(void);
    extern int64_t  PPIDD_Worker_comm(void);
    extern void PPIDD_Finalize(void);
